+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 17 Mar 2021
 
 ### Added
+ngx-slice-kit:
+- background and color default body variables
 
 ### Changed
 
 ### Fixed
+ngx-slice-kit:
 - public api imports to prevent errors like:
   `Unsupported private class LoadingModule. This class is visible to consumers via SliceKitModule -> LayoutModule -> LoadingModule, but is not exported from the top-level library entrypoint.`